Move TransactionEnv conversion to From<(Transaction, Address)> impl in mod.rs, next to the proto type definitions. Removes build_transaction_env from client.rs — the handler now just calls (tx, sender).into().
